+++ tombio/score.js
@@ -16,6 +16,8 @@
 
         //Only scores with values are passed in, so no need to deal with missing values.
 
+        //console.log(stateval, rng, wholeRange, kbStrictness);
+
         var scorefor;
         var latitude = (1 - kbStrictness / 10) * wholeRange;
         if (stateval >= rng.min && stateval <= rng.max) {
@@ -37,11 +39,6 @@
     }
 
     exports.ordinal = function (selState, kbTaxonStates, posStates, kbStrictness) {
-
-        console.log("selState", selState)
-        console.log("kbTaxonStates", kbTaxonStates)
-        console.log("posStates", posStates)
-        console.log("kbStrictness", kbStrictness)
 
         //selState is the state we're assessing for a match.
         //kbTaxonStates are the states recorded in the KB for the taxon (already adjusted for sex).
@@ -77,37 +74,74 @@
         return exports.numberVsRange(stateval, rng, posStates.length - 1, kbStrictness)
     }
 
-    exports.ordinal2 = function (selStates, kbTaxonStates, posStates, kbStrictness, isCircular) {
+    exports.ordinal2 = function (selectedStates, kbTaxonStates, posStates, kbStrictness, isCircular) {
 
-        console.log("selStates", selStates)
-        console.log("kbTaxonStates", kbTaxonStates)
-        console.log("posStates", posStates)
-        console.log("kbStrictness", kbStrictness)
-        console.log("isCircular", isCircular)
+        //console.log("selectedStates", selectedStates)
+        //console.log("kbTaxonStates", kbTaxonStates)
+        //console.log("posStates", posStates)
+        //console.log("kbStrictness", kbStrictness)
+        //console.log("isCircular", isCircular)
+
+        //selectedStates is an array of states we're assessing for a match. Can be more than one for multiple selection controls.
+        //kbTaxonStates are the states recorded in the KB for the taxon (already adjusted for sex) - note
+        //that any of these states could be expressed as ordinal ranges so this parameter is an array of
+        //arrays. Within the outer array, single ordinals are an array of size one. In this routine, all the arrays within
+        //the outer array will simply be merged into a single array and all treated as alternatives.
+        //posStates are the ordinal states that can be taken on by this character.
+        //kbStrictness is the strictness for this character in the KB.
+
+        //This function changes all the ordinal states into scores and then simply refers
+        //the resulting numbers to the numberVsRange function.
 
         //Deal first with missing KB values
         if (kbTaxonStates.length == 0) {
             return [0, 0];
         }
 
-        return [0, 0];
+        //Merge the kbTaxonStates arrays.
+        var kbAllTaxonStates = [];
+        kbTaxonStates.forEach(function (a) {
+            kbAllTaxonStates = kbAllTaxonStates.concat(a);
+        });
 
-        //var stateval;
-        //var rng = { min: null, max: null }
-        //posStates.forEach(function (state, rank) {
-        //    if (state == selState) {
-        //        stateval = rank;
-        //    }
-        //    kbTaxonStates.forEach(function (taxState) {
-        //        if (!rng.min && taxState == state) {
-        //            rng.min = rank;
-        //        }
-        //        if (taxState == state) {
-        //            rng.max = rank;
-        //        }
-        //    })
-        //})
-        //return exports.numberVsRange(stateval, rng, posStates.length - 1, kbStrictness)
+        //Return the best score
+        var ret;
+        selectedStates.forEach(function (selState) {    
+            var selStateRank;
+            posStates.forEach(function (state, rank) {
+                if (state == selState) selStateRank = rank;
+            })
+            kbAllTaxonStates.forEach(function (taxState) {
+                var taxStateRank;
+                posStates.forEach(function (state, rank) {
+                    if (state == taxState) taxStateRank = rank;
+                })
+                var rngTaxonStateRank = { min: taxStateRank, max: taxStateRank };
+                var r = exports.numberVsRange(selStateRank, rngTaxonStateRank, posStates.length - 1, kbStrictness);
+                //console.log("standard", r[0])
+                if (!ret || ret[0] < r[0]) {
+                    ret = r;
+                }
+                //If it is a circular ordinal then change the value of taxStateRank by either subtracting or adding
+                //the number of all values - whichever brings the value of taxStateRank and selStateRank closer
+                //together - and score for those values. If the score is better than the unmodified taxStateRank then
+                //use it instead.
+                if (isCircular) {
+                    if (selStateRank > taxStateRank) {
+                        taxStateRank = taxStateRank + posStates.length;
+                    } else {
+                        taxStateRank = taxStateRank - posStates.length;
+                    }
+                    var rngTaxonStateRankC = { min: taxStateRank, max: taxStateRank};
+                    var rC = exports.numberVsRange(selStateRank, rngTaxonStateRankC, posStates.length - 1, kbStrictness);
+                    //console.log("circular", rC[0])
+                    if (!ret || ret[0] < rC[0]) {
+                        ret = rC; 
+                    }
+                }
+            })
+        })
+        return ret;
     }
 
     exports.character = function (selectedStates, kbTaxonStates) {